Move the color module to src/node
