+++ configurations/tsconfig.json
@@ -20,6 +20,7 @@
         "@shopify/cli-hydrogen/commands/hydrogen/init": ["../packages/cli-hydrogen/src/cli/commands/hydrogen/init.ts"],
         "@shopify/cli-kit": ["../packages/cli-kit/src/index.ts"],
         "@shopify/cli-kit/node/*": ["../packages/cli-kit/src/node/*.ts"],
+        "@shopify/cli-kit/common/*": ["../packages/cli-kit/src/common/*.ts"],
       }
     },
     "exclude": ["**/dist/**"]
