+++ packages/cli-kit/package.json
@@ -21,6 +21,10 @@
       "node": "./dist/node/*.js",
       "types": "./dist/node/*.d.ts"
     },
+    "./common/*": {
+      "node": "./dist/common/*.js",
+      "types": "./dist/common/*.d.ts"
+    },
     "./testing/*": {
       "node": "./dist/testing/*.js",
       "types": "./dist/testing/*.d.ts"
