+++ packages/cli-kit/src/output.ts
@@ -13,7 +13,7 @@ import {
 } from './file.js'
 import {join as pathJoin, relativize as relativizePath} from './path.js'
 import {page} from './system.js'
-import {colors} from './colors.js'
+import {colors} from './node/colors.js'
 import terminalLink from 'terminal-link'
 import StackTracey from 'stacktracey'
 import {AbortController, AbortSignal} from 'abort-controller'
